requirements.txt
Flask==2.0.1
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
